example/requirements.txt
-r ../requirements/dist.txt

factory_boy
django-extensions
